+++ packages/cli/create-medusa-app/src/utils/nextjs-utils.ts
@@ -7,8 +7,8 @@ import { isAbortError } from "./create-abort-controller.js"
 import execute from "./execute.js"
 import { displayFactBox, FactBoxOptions } from "./facts.js"
 import logMessage from "./log-message.js"
-import ProcessManager from "./process-manager.js"
 import { updatePackageVersions } from "./update-package-versions.js"
+import PackageManager from "./package-manager.js"
 
 const NEXTJS_REPO = "https://github.com/medusajs/nextjs-starter-medusa"
 const NEXTJS_BRANCH = "main"
@@ -31,7 +31,7 @@ type InstallOptions = {
   abortController?: AbortController
   factBoxOptions: FactBoxOptions
   verbose?: boolean
-  processManager: ProcessManager
+  packageManager: PackageManager
   version?: string
 }
 
@@ -40,7 +40,7 @@ export async function installNextjsStarter({
   abortController,
   factBoxOptions,
   verbose = false,
-  processManager,
+  packageManager,
   version,
 }: InstallOptions): Promise<string> {
   factBoxOptions.interval = displayFactBox({
@@ -83,20 +83,7 @@ export async function installNextjsStarter({
       signal: abortController?.signal,
       cwd: nextjsDirectory,
     }
-    await processManager.runProcess({
-      process: async () => {
-        try {
-          await execute([`yarn`, execOptions], { verbose })
-        } catch (e) {
-          // yarn isn't available
-          // use npm
-          await execute([`npm install`, execOptions], {
-            verbose,
-          })
-        }
-      },
-      ignoreERESOLVE: true,
-    })
+    await packageManager.installDependencies(execOptions)
   } catch (e) {
     if (isAbortError(e)) {
       process.exit()
@@ -130,14 +117,17 @@ type StartOptions = {
   directory: string
   abortController?: AbortController
   verbose?: boolean
+  packageManager: PackageManager
 }
 
 export function startNextjsStarter({
   directory,
   abortController,
   verbose = false,
+  packageManager,
 }: StartOptions) {
-  const childProcess = exec(`npm run dev`, {
+  const command = packageManager.getCommandStr(`dev`)
+  const childProcess = exec(command, {
     cwd: directory,
     signal: abortController?.signal,
   })