+++ packages/core/core-flows/src/translation/steps/create-translations.ts
@@ -5,23 +5,31 @@ import {
 import { Modules } from "@medusajs/framework/utils"
 import { StepResponse, createStep } from "@medusajs/framework/workflows-sdk"
 
+/**
+ * The translations to create.
+ */
+export type CreateTranslationsStepInput = CreateTranslationDTO[]
+
 export const createTranslationsStepId = "create-translations"
 /**
  * This step creates one or more translations.
+ * 
+ * @since 2.12.3
+ * @featureFlag translation
  *
  * @example
  * const data = createTranslationsStep([
  *   {
  *     reference_id: "prod_123",
  *     reference: "product",
- *     locale_code: "fr-FR",
+ *     locale: "fr-FR",
  *     translations: { title: "Produit", description: "Description du produit" }
  *   }
  * ])
  */
 export const createTranslationsStep = createStep(
   createTranslationsStepId,
-  async (data: CreateTranslationDTO[], { container }) => {
+  async (data: CreateTranslationsStepInput, { container }) => {
     const service = container.resolve<ITranslationModuleService>(
       Modules.TRANSLATION
     )