+++ packages/core/framework/src/medusa-app-loader.ts
@@ -157,17 +157,15 @@ export class MedusaAppLoader {
    * @param action
    */
   async runModulesMigrations(
-    {
-      moduleNames,
-      action = "run",
-    }:
+    options:
       | {
-          moduleNames?: never
           action: "run"
+          allOrNothing?: boolean
         }
       | {
-          moduleNames: string[]
           action: "revert" | "generate"
+          moduleNames: string[]
+          allOrNothing?: never
         } = {
       action: "run",
     }
@@ -185,14 +183,15 @@ export class MedusaAppLoader {
       injectedDependencies,
       medusaConfigPath: this.#medusaConfigPath,
       cwd: this.#cwd,
+      allOrNothing: options.allOrNothing,
     }
 
-    if (action === "revert") {
-      await MedusaAppMigrateDown(moduleNames!, migrationOptions)
-    } else if (action === "run") {
+    if (options.action === "revert") {
+      await MedusaAppMigrateDown(options.moduleNames!, migrationOptions)
+    } else if (options.action === "run") {
       await MedusaAppMigrateUp(migrationOptions)
-    } else {
-      await MedusaAppMigrateGenerate(moduleNames!, migrationOptions)
+    } else if (options.action === "generate") {
+      await MedusaAppMigrateGenerate(options.moduleNames!, migrationOptions)
     }
   }
 
@@ -311,13 +310,25 @@ export class MedusaAppLoader {
    * Load all modules and bootstrap all the modules and links to be ready to be consumed
    * @param config
    */
-  async load(config = { registerInContainer: true }): Promise<MedusaAppOutput> {
+  async load(
+    config: {
+      registerInContainer?: boolean
+      schemaOnly?: boolean
+      migrationOnly?: boolean
+    } = {
+      registerInContainer: true,
+      schemaOnly: false,
+      migrationOnly: false,
+    }
+  ): Promise<MedusaAppOutput> {
     const configModule: ConfigModule = this.#container.resolve(
       ContainerRegistrationKeys.CONFIG_MODULE
     )
 
     const { sharedResourcesConfig, injectedDependencies } =
-      this.prepareSharedResourcesAndDeps()
+      !config.migrationOnly && !config.schemaOnly
+        ? this.prepareSharedResourcesAndDeps()
+        : {}
 
     this.#container.register(
       ContainerRegistrationKeys.REMOTE_QUERY,
@@ -344,6 +355,8 @@ export class MedusaAppLoader {
       injectedDependencies,
       medusaConfigPath: this.#medusaConfigPath,
       cwd: this.#cwd,
+      migrationOnly: config.migrationOnly,
+      schemaOnly: config.schemaOnly,
     })
 
     if (!config.registerInContainer) {