+++ packages/core/modules-sdk/src/medusa-module.ts
@@ -90,6 +90,8 @@ export type LinkModuleBootstrapOptions = {
   moduleExports?: ModuleExports
   injectedDependencies?: Record<string, any>
   cwd?: string
+  migrationOnly?: boolean
+  schemaOnly?: boolean
 }
 
 export type RegisterModuleJoinerConfig =
@@ -300,18 +302,20 @@ class MedusaModule {
   public static async bootstrapAll(
     modulesOptions: Omit<
       ModuleBootstrapOptions,
-      "migrationOnly" | "loaderOnly" | "workerMode"
+      "migrationOnly" | "loaderOnly" | "workerMode" | "schemaOnly"
     >[],
     {
       migrationOnly,
       loaderOnly,
       workerMode,
+      schemaOnly,
       cwd,
     }: {
       migrationOnly?: boolean
       loaderOnly?: boolean
       workerMode?: ModuleBootstrapOptions["workerMode"]
       cwd?: string
+      schemaOnly?: boolean
     }
   ): Promise<
     {
@@ -323,6 +327,7 @@ class MedusaModule {
       loaderOnly,
       workerMode,
       cwd,
+      schemaOnly,
     })
   }
 
@@ -391,18 +396,20 @@ class MedusaModule {
   protected static async bootstrap_<T>(
     modulesOptions: Omit<
       ModuleBootstrapOptions,
-      "migrationOnly" | "loaderOnly" | "workerMode" | "cwd"
+      "migrationOnly" | "loaderOnly" | "workerMode" | "cwd" | "schemaOnly"
     >[],
     {
       migrationOnly,
       loaderOnly,
       workerMode,
       cwd = process.cwd(),
+      schemaOnly,
     }: {
       migrationOnly?: boolean
       loaderOnly?: boolean
       workerMode?: "shared" | "worker" | "server"
       cwd?: string
+      schemaOnly?: boolean
     }
   ): Promise<
     {
@@ -507,6 +514,7 @@ class MedusaModule {
             moduleResolutions,
             logger: logger_,
             migrationOnly,
+            schemaOnly,
             loaderOnly,
           })
         } catch (err) {
@@ -652,6 +660,8 @@ class MedusaModule {
     moduleExports,
     injectedDependencies,
     cwd,
+    migrationOnly,
+    schemaOnly,
   }: LinkModuleBootstrapOptions): Promise<{
     [key: string]: unknown
   }> {
@@ -720,6 +730,8 @@ class MedusaModule {
       await moduleLoader({
         container,
         moduleResolutions,
+        migrationOnly,
+        schemaOnly,
         logger: logger_,
       })
     } catch (err) {
@@ -828,7 +840,7 @@ class MedusaModule {
     moduleKey,
     modulePath,
     cwd,
-  }: MigrationOptions): Promise<void> {
+  }: MigrationOptions): Promise<{ name: string; path: string }[]> {
     const moduleResolutions = registerMedusaModule({
       moduleKey,
       moduleDeclaration: {
@@ -846,6 +858,7 @@ class MedusaModule {
 
     container ??= createMedusaContainer()
 
+    let result: { name: string; path: string }[] = []
     for (const mod in moduleResolutions) {
       const { runMigrations } = await loadModuleMigrations(
         container,
@@ -854,23 +867,29 @@ class MedusaModule {
       )
 
       if (typeof runMigrations === "function") {
-        await runMigrations({
+        const res = await runMigrations({
           options,
           container: container!,
           logger: logger_,
         })
+        result.push(...res)
       }
     }
+
+    return result
   }
 
-  public static async migrateDown({
-    options,
-    container,
-    moduleExports,
-    moduleKey,
-    modulePath,
-    cwd,
-  }: MigrationOptions): Promise<void> {
+  public static async migrateDown(
+    {
+      options,
+      container,
+      moduleExports,
+      moduleKey,
+      modulePath,
+      cwd,
+    }: MigrationOptions,
+    migrationNames?: string[]
+  ): Promise<void> {
     const moduleResolutions = registerMedusaModule({
       moduleKey,
       moduleDeclaration: {
@@ -900,6 +919,7 @@ class MedusaModule {
           options,
           container: container!,
           logger: logger_,
+          migrationNames,
         })
       }
     }