+++ packages/medusa/src/api/admin/locales/[code]/route.ts
@@ -1,7 +1,17 @@
-import { ContainerRegistrationKeys } from "@medusajs/framework/utils"
+import {
+  ContainerRegistrationKeys,
+  defineFileConfig,
+  FeatureFlag,
+  MedusaError,
+} from "@medusajs/framework/utils"
 import { MedusaRequest, MedusaResponse } from "@medusajs/framework/http"
 import { HttpTypes } from "@medusajs/framework/types"
+import TranslationFeatureFlag from "../../../../feature-flags/translation"
 
+/**
+ * @since 2.12.3
+ * @featureFlag translation
+ */
 export const GET = async (
   req: MedusaRequest<HttpTypes.AdminLocaleParams>,
   res: MedusaResponse<HttpTypes.AdminLocaleResponse>
@@ -20,9 +30,19 @@ export const GET = async (
     },
     {
       cache: { enable: true },
-      throwIfKeyNotFound: true,
     }
   )
 
+  if (!locale) {
+    throw new MedusaError(
+      MedusaError.Types.NOT_FOUND,
+      `Locale with code: ${req.params.code} was not found`
+    )
+  }
+
   res.status(200).json({ locale })
 }
+
+defineFileConfig({
+  isDisabled: () => !FeatureFlag.isFeatureEnabled(TranslationFeatureFlag.key),
+})