+++ packages/medusa/src/api/admin/translations/batch/route.ts
@@ -6,13 +6,17 @@ import {
   FeatureFlag,
 } from "@medusajs/framework/utils"
 import { BatchMethodRequest, HttpTypes } from "@medusajs/types"
+import TranslationFeatureFlag from "../../../../feature-flags/translation"
 import { defaultAdminTranslationFields } from "../query-config"
 import {
   AdminCreateTranslationType,
   AdminUpdateTranslationType,
 } from "../validators"
-import TranslationFeatureFlag from "../../../../feature-flags/translation"
 
+/**
+ * @since 2.12.3
+ * @featureFlag translation
+ */
 export const POST = async (
   req: AuthenticatedMedusaRequest<
     BatchMethodRequest<AdminCreateTranslationType, AdminUpdateTranslationType>